+++ src/context/GitClientContext.tsx
@@ -371,6 +371,7 @@ interface GitClientContextType {
   commitMsg: string;
   setCommitMsg: (msg: string) => void;
   commits: CommitRaw[];
+  totalCommitCount: number;
   getCommitHash: (i: number) => string;
   getCommitFullSha: (i: number) => string;
   graphData: GraphData;